requirements.txt
torch==1.10.2
torchvision
numpy==1.22.2
transformers==4.16.0
git+https://github.com/openai/CLIP.git
tqdm
Pillow
python-multipart
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0